+++ Classes/ZoomScriptScene.h
@@ -91,8 +91,8 @@ private:
 	
 	bool isAnimated;
 	CCPoint touch_p;
-    long touchStartTime;
-	long first_touch_time;
+    unsigned long long touchStartTime;
+	unsigned long long first_touch_time;
 	int first_touch_p;
 	CCPoint first_touch_point;
 	CCPoint touchStart_p;
